Corrected Air Base locations Southern Virginia and some in Germany
Air Bases we too bunched up or in unrealistic locations. German Air bases many were moved to locations in the state where there was a high concentration or in the middle of the major air bases.
Southern Virginia air base was moved to Norfolk to represent the Naval Air Station Oceana, which played a huge part in American naval patrols in the Atlantic.