Make Ship Supply Irrelevant (Define version)
Ship supply (doesn't) work in current HoI4 by having ships draw all supply from the closest friendly port, no matter how small it may be or how large a nearby other port may be.
In order to make it so that this does not fuck over the Pacific and Atlantic theatres, ship supply usage made so low as to be irrelevant.
Done using defines, so easier to revert if needed