bumped AOT penalty further
Last bump from 10x to 15x barely did anything due to the effect reduction the closer you come to the actual year. The difference between 10x and 15x only increased the research time when researching 0,5 years AOT by just a couple days. not the desired effect. Bumping it up to 25x results in the desired effect