US War Focuses Rework
Prepare for War replaces War Plans Division. Adds Develop the Rainbow Plans. This consumes political power and reduces command power gain for a period, after which it gives the bonuses that were originally tied to the Army of the United States and Department of Defence focuses, both of which have been removed.

Fortify the Coast is an entirely new focus. It gives access to a decision to build radar and state-based AA across the US east and west coasts, to represent the construction that was done.

Fortify the Pacific Territories replaces the original Defend the Pacific that was tied to the War Plan against Japan. Instead of giving naval bases and coastal forts for free, it now works as a standard construction decision.

Build the Pentagon has been changed to be actually building the Pentagon. It gives a decision that starts a chain of missions, consuming a small number of civs over months where every 98 days a new wedge is completed giving more of a bonus.

Louisiana Maneuvers has been updated to function a little closer to Germany's Lessons from Poland decision. Once you have a certain amount of manpower deployed, you can take the decision at a cost of army xp and a little bit of PP, plus the consumption of CP over time. After the decision expires, you get 2 doctrines for free.

Destroyers for Bases now no longer requests Iceland, since that remained in British hands during the war.

All of the original war plan focuses have been replaced with a focus each for 4 major US operations. Operation Torch is untouched. Operation Overlord combines the current Final Victory propaganda with War Plan Black. Operation Downfall combines the current Final Victory propaganda with War Plan Orange, plus 2 additional naval invasion cap. Operation Unthinkable gives the US a war goal against the Soviet Union, plus the original War Plan White bonus. As part of this, the Final Victory propaganda has been removed. Moreover, Downfall and Overlord are mutually exclusive.